+++ ferry_store/lib/src/memory_store.dart
@@ -1,46 +1,38 @@
 import 'dart:async';
+import 'package:rxdart/rxdart.dart';
 
 import './store.dart';
 
 class MemoryStore extends Store {
-  final _controller = StreamController<Map<String, Map<String, dynamic>>>();
+  final BehaviorSubject<Map<String, Map<String, dynamic>>> _dataController;
 
-  final Map<String, Map<String, dynamic>> _data;
-
-  MemoryStore([Map<String, Map<String, dynamic>> data]) : _data = data ?? {} {
-    _controller.add(_data);
-  }
+  MemoryStore([Map<String, Map<String, dynamic>> initialData])
+      : _dataController = BehaviorSubject.seeded(initialData ?? {});
 
   @override
   Stream<Map<String, Map<String, dynamic>>> watch() {
-    return _controller.stream;
+    return _dataController.stream;
   }
 
   @override
   Map<String, dynamic> get(String dataId) {
-    return _data[dataId];
+    return _dataController.value[dataId];
   }
 
   @override
-  void put(String dataId, Map<String, dynamic> value) {
-    _data[dataId] = value;
-    _controller.add(_data);
-  }
+  void put(String dataId, Map<String, dynamic> value) =>
+      _dataController.add(_dataController.value..[dataId] = value);
 
   @override
-  void putAll(Map<String, Map<String, dynamic>> data) {
-    _data.addAll(data);
-    _controller.add(_data);
-  }
+  void putAll(Map<String, Map<String, dynamic>> data) =>
+      _dataController.add(_dataController.value..addAll(data));
 
   @override
-  void delete(String dataId) {
-    _data.remove(dataId);
-    _controller.add(_data);
-  }
+  void delete(String dataId) =>
+      _dataController.add(_dataController.value..remove(dataId));
 
   @override
-  Map<String, Map<String, dynamic>> toMap() {
-    return {..._data};
-  }
+  Map<String, Map<String, dynamic>> toMap() => Map.from(_dataController.value);
+
+  Future<void> dispose() => _dataController.close();
 }